# backend/specs/006-program-batches/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Program Batches API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - accessCookie: []
paths:
...
  /batches/{batchId}/eligibility:
    get:
      summary: Evaluate Admissions eligibility
      parameters:
        - { $ref: '#/components/parameters/BatchId' }
        - { name: branchId, in: query, required: true, schema: { type: string, format: uuid } }
        - { name: today, in: query, required: true, schema: { type: string, format: date } }
      responses:
        '200': { description: Eligibility envelope }
        '503': { $ref: '#/components/responses/Error' }
  /batches/{batchId}/lifecycle:
    get:
      summary: Read immutable lifecycle history
      parameters: [{ $ref: '#/components/parameters/BatchId' }]
      responses:
        '200': { description: Lifecycle envelope }
  /batches/{batchId}/financial-revisions:
    get:
      summary: Read immutable financial revisions
      parameters: [{ $ref: '#/components/parameters/BatchId' }]
      responses:
        '200': { description: Financial revision envelope }
components:
  securitySchemes:
    accessCookie: { type: apiKey, in: cookie, name: access_token }
  parameters:
...
    BatchId: { name: batchId, in: path, required: true, schema: { type: string, format: uuid } }
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error: { type: object, required: [code, message] }
  responses:
...
    Error:
      description: Closed Arabic error envelope
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }